Use Query Builder for better support across platforms.
Not all drivers supported FOR UPDATE queries so this changes changes the
raw SQL to use the query builder thus allowing DBAL to construct hte
appropriate query per platform. It then grabs the SQL and appends the
read lock SQL but grabs the SQL part from the correct platform thus as
to not break portability.